重组为石大时光圈，添加UUID依赖，重构事件模型，优化主题常量，更新事件项组件，调整文件结构，删除冗余代码
