+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   html: ^0.15.0
   path_provider: ^2.0.11
   webview_flutter: ^4.10.0
+  uuid: ^4.5.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
